+++ src/Mtlstats/Menu.hs
@@ -27,7 +27,8 @@ module Mtlstats.Menu (
   mainMenu,
   newSeasonMenu,
   gameMonthMenu,
-  gameTypeMenu
+  gameTypeMenu,
+  editPlayerMenu
 ) where
 
 import Control.Monad.IO.Class (liftIO)
@@ -73,7 +74,9 @@ mainMenu = Menu "*** MAIN MENU ***" True
     modify createPlayer >> return True
   , MenuItem '4' "Create Goalie" $
     modify createGoalie >> return True
-  , MenuItem '5' "Exit" $ do
+  , MenuItem '5' "Edit Player" $
+    modify editPlayer >> return True
+  , MenuItem '6' "Exit" $ do
     db <- gets $ view database
     liftIO $ do
       dir <- getAppUserDataDirectory appName
@@ -120,4 +123,22 @@ gameTypeMenu = Menu "Game type:" ()
     modify $ progMode.gameStateL.gameType ?~ HomeGame
   , MenuItem '2' "Away Game" $
     modify $ progMode.gameStateL.gameType ?~ AwayGame
+  ]
+
+-- | The player edit menu
+editPlayerMenu :: Menu ()
+editPlayerMenu = Menu "*** EDIT PLAYER ***" () $ map
+  (\(ch, label, mode) -> MenuItem ch label $ case mode of
+    Nothing -> modify $ progMode .~ MainMenu
+    Just m  -> modify $ progMode.editPlayerStateL.epsMode .~ m)
+  [ ( '1', "Change number",         Just EPNumber     )
+  , ( '2', "Change name",           Just EPName       )
+  , ( '3', "Change position",       Just EPPosition   )
+  , ( '4', "YTD goals",             Just EPYtdGoals   )
+  , ( '5', "YTD assists",           Just EPYtdAssists )
+  , ( '6', "YTD penalty mins",      Just EPYtdPMin    )
+  , ( '7', "Lifetime goals",        Just EPLtGoals    )
+  , ( '8', "Lifetime assists",      Just EPLtAssists  )
+  , ( '9', "Lifetime penalty mins", Just EPLtPMin     )
+  , ( '0', "Finished editing",      Nothing           )
   ]